# packages/api-client/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Botonera2 Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/preparacion": {
      "post": {
        "tags": [
          "preparacion"
        ],
        "summary": "Preparar Sala",
        "description": "Inicia una nueva preparación del recinto (CU-01).\n\nCarga y congela ``config/system.toml`` y ``config/concejales.csv``, crea\nel conjunto de auditoría L1/L2/L3, persiste ``PREPARACION_INICIADA`` y\nrecién entonces deja el sistema en ``PREPARANDO`` con todos los\nconcejales ausentes. Si algún paso obligatorio falla, no se confirma\néxito y el sistema permanece en ``SIN_PREPARAR``.",
        "operationId": "preparar_sala_api_v1_preparacion_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "409": {
            "description": "El comando no es válido para el estado global actual (ESTADO_INCOMPATIBLE).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "503": {
            "description": "Indisponibilidad técnica: CONFIGURACION_INVALIDA, PADRON_INVALIDO o AUDITORIA_NO_DISPONIBLE según el caso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "500": {
            "description": "Fallo inesperado no clasificado (ERROR_INTERNO).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "preparacion"
        ],
        "summary": "Cancelar Preparacion",
        "description": "Cancela la preparación activa (CU-02), sin recibir ni exigir motivo.\n\nPersiste ``PREPARACION_CANCELADA``, cierra definitivamente los tres CSV\n(que se conservan en disco) y devuelve el sistema a ``SIN_PREPARAR`` con\nel contexto operativo completamente descartado. Si la auditoría no puede\ngarantizar esos pasos, no se confirma la cancelación y el estado\npermanece en ``PREPARANDO`` en fallo cerrado.",
        "operationId": "cancelar_preparacion_api_v1_preparacion_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "409": {
            "description": "El comando no es válido para el estado global actual (ESTADO_INCOMPATIBLE).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "503": {
            "description": "Indisponibilidad técnica: CONFIGURACION_INVALIDA, PADRON_INVALIDO o AUDITORIA_NO_DISPONIBLE según el caso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "500": {
            "description": "Fallo inesperado no clasificado (ERROR_INTERNO).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "preparacion"
        ],
        "summary": "Actualizar número y autoridades de la preparación",
        "description": "Aplica uno o más cambios efectivos con auditoría previa y ordenada.",
        "operationId": "actualizar_preparacion_api_v1_preparacion_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitudActualizarPreparacion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "409": {
            "description": "El comando no es válido para el estado global actual (ESTADO_INCOMPATIBLE).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "503": {
            "description": "Indisponibilidad técnica: CONFIGURACION_INVALIDA, PADRON_INVALIDO o AUDITORIA_NO_DISPONIBLE según el caso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "500": {
            "description": "Fallo inesperado no clasificado (ERROR_INTERNO).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Body vacío/inválido o número que no es entero estricto positivo."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorRespuesta": {
        "properties": {
          "codigo": {
            "type": "string",
            "title": "Codigo"
          },
          "mensaje": {
            "type": "string",
            "title": "Mensaje"
          }
        },
        "type": "object",
        "required": [
          "codigo",
          "mensaje"
        ],
        "title": "ErrorRespuesta",
        "description": "Cuerpo JSON estable de toda respuesta de error funcional/técnica.\n\n``codigo`` es el identificador estable legible por máquina; ``mensaje``\naporta diagnóstico humano seguro (nunca trazas ni datos internos)."
      },
...
      "NumeroSesionOmitible": {
        "type": "integer",
        "minimum": 1.0
      },
...
      "SolicitudActualizarPreparacion": {
        "properties": {
          "numero_sesion": {
            "$ref": "#/components/schemas/NumeroSesionOmitible"
          },
          "presidencia": {
            "$ref": "#/components/schemas/TextoOmitible"
          },
          "secretaria_legislativa": {
            "$ref": "#/components/schemas/TextoOmitible"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SolicitudActualizarPreparacion",
        "description": "Body parcial para número y autoridades durante ``PREPARANDO``.\n\nCada propiedad puede omitirse, pero el body debe contener al menos una.\nLos textos vacíos son válidos aquí porque significan limpiar la autoridad.\nEl entero estricto evita la coerción de booleanos, floats y strings."
      },
...
      "TextoOmitible": {
        "type": "string"
      },
...
    }
  }
}